Refactor: stats Aggregator の apply_*_filter を FilterableConcern に集約
14 の Stats Aggregator / Service が apply_year_filter / apply_match_type_filter
/ apply_season_filter / apply_tournament_filter の 4 メソッドを完全に同一
実装で重複保持していた。フィルタ仕様変更時に 14 箇所を手で揃える必要が
あり、本 PR の前段 (#371) でも 14 ファイルすべての apply_year_filter を
触らざるを得なかった。

Stats::Concerns::FilterableConcern に 4 メソッドを集約し、14 Aggregator
を include に差し替える。include 側は @year / @match_type / @season_id /
@tournament_id を initialize で必ずセットしている前提（既存の 14 ファイル
すべてで満たされている）。

apply_year_filter は #371 で Time.zone ベースに統一済みなので、本 PR では
その実装をそのまま Concern に移すだけ。EraTrendService / table service 系
は filter シグネチャが部分的に異なる（match_type を持たない / season /
tournament の扱いが違う）ため対象外。
