+++ app/services/stats/additional_stats_aggregator.rb
@@ -10,6 +10,8 @@ module Stats
   # フィルタは HeadlineStatsAggregator と同じ 5 項目（year / match_type /
   # season_id / tournament_id）を踏襲する。
   class AdditionalStatsAggregator
+    include Concerns::FilterableConcern
+
     SUM_COLUMNS = %i[
       plate_appearances at_bats hit total_bases
       two_base_hit three_base_hit
@@ -85,34 +87,6 @@ module Stats
         scope = apply_season_filter(scope)
         apply_tournament_filter(scope)
       end
-    end
-
-    def apply_year_filter(scope)
-      return scope if @year.blank? || @year.to_s == '通算'
-
-      yr = @year.to_i
-      range_start = Time.zone.local(yr, 1, 1)
-      range_end = Time.zone.local(yr + 1, 1, 1)
-      scope.where('match_results.date_and_time >= ? AND match_results.date_and_time < ?',
-                  range_start, range_end)
-    end
-
-    def apply_match_type_filter(scope)
-      return scope if @match_type.blank? || @match_type == '全て'
-
-      scope.where(match_results: { match_type: @match_type })
-    end
-
-    def apply_season_filter(scope)
-      return scope if @season_id.blank?
-
-      scope.where(game_results: { season_id: @season_id })
-    end
-
-    def apply_tournament_filter(scope)
-      return scope if @tournament_id.blank?
-
-      scope.where(match_results: { tournament_id: @tournament_id })
     end
 
     def safe_divide(numerator, denominator)
