+++ app/services/stats/batting_trend_aggregator.rb
@@ -11,6 +11,8 @@ module Stats
   # granularity=game は **累積** を返す（最初の試合から各時点までの通算成績）。
   # granularity=month は **その月単独** を返す（各月でリセット）。
   class BattingTrendAggregator # rubocop:disable Metrics/ClassLength
+    include Concerns::FilterableConcern
+
     SUM_COLUMNS = %i[at_bats hit total_bases base_on_balls hit_by_pitch sacrifice_fly].freeze
 
     # 受け付ける granularity:
@@ -196,34 +198,6 @@ module Stats
         scope = apply_season_filter(scope)
         apply_tournament_filter(scope)
       end
-    end
-
-    def apply_year_filter(scope)
-      return scope if @year.blank? || @year.to_s == '通算'
-
-      yr = @year.to_i
-      range_start = Time.zone.local(yr, 1, 1)
-      range_end = Time.zone.local(yr + 1, 1, 1)
-      scope.where('match_results.date_and_time >= ? AND match_results.date_and_time < ?',
-                  range_start, range_end)
-    end
-
-    def apply_match_type_filter(scope)
-      return scope if @match_type.blank? || @match_type == '全て'
-
-      scope.where(match_results: { match_type: @match_type })
-    end
-
-    def apply_season_filter(scope)
-      return scope if @season_id.blank?
-
-      scope.where(game_results: { season_id: @season_id })
-    end
-
-    def apply_tournament_filter(scope)
-      return scope if @tournament_id.blank?
-
-      scope.where(match_results: { tournament_id: @tournament_id })
     end
 
     def safe_divide(numerator, denominator)
