+++ app/services/stats/plate_appearance_breakdown_service.rb
@@ -2,6 +2,8 @@
 
 module Stats
   class PlateAppearanceBreakdownService
+    include Concerns::FilterableConcern
+
     CATEGORIES = {
       '単打' => [7],
       '長打' => [8, 9],
@@ -45,34 +47,6 @@ module Stats
       scope = apply_match_type_filter(scope)
       scope = apply_season_filter(scope)
       apply_tournament_filter(scope)
-    end
-
-    def apply_year_filter(scope)
-      return scope if @year.blank? || @year.to_s == '通算'
-
-      yr = @year.to_i
-      range_start = Time.zone.local(yr, 1, 1)
-      range_end = Time.zone.local(yr + 1, 1, 1)
-      scope.where('match_results.date_and_time >= ? AND match_results.date_and_time < ?',
-                  range_start, range_end)
-    end
-
-    def apply_match_type_filter(scope)
-      return scope if @match_type.blank? || @match_type == '全て'
-
-      scope.where(match_results: { match_type: @match_type })
-    end
-
-    def apply_season_filter(scope)
-      return scope if @season_id.blank?
-
-      scope.where(game_results: { season_id: @season_id })
-    end
-
-    def apply_tournament_filter(scope)
-      return scope if @tournament_id.blank?
-
-      scope.where(match_results: { tournament_id: @tournament_id })
     end
   end
 end
